Fixed rotation worm

+++ src/banana_missile.cpp
@@ -3,8 +3,8 @@
 
 BananaMissile::BananaMissile(Stage& stage
 	, Worm* owner
-	, const int x
-	, const int y
+	, const float x
+	, const float y
 	, const float angle_rad
 	, const b2Vec2 velocity
 	, const float angular_velocity
