+++ src/bazooka_missile.h
@@ -9,8 +9,8 @@ class BazookaMissile : public Throwable {
 public:
 	BazookaMissile(Stage& stage
 	, Worm* owner
-	, const int x
-	, const int y
+	, const float x
+	, const float y
 	, const float angle_rad
 	, const b2Vec2 velocity
 	, const float angular_velocity
