+++ src/delayed_throwable.cpp
@@ -4,8 +4,8 @@
 
 DelayedThrowable::DelayedThrowable(Stage& stage
 	, Worm* owner
-	, const int x
-	, const int y
+	, const float x
+	, const float y
 	, const float angle_rad
 	, const b2Vec2 velocity
 	, const float angular_velocity
