+++ src/delayed_throwable.h
@@ -13,8 +13,8 @@ protected:
 public:
 	DelayedThrowable(Stage& stage
 	, Worm* owner	
-	, const int x
-	, const int y
+	, const float x
+	, const float y
 	, const float angle_rad
 	, const b2Vec2 velocity
 	, const float angular_velocity
