+++ src/throwable.h
@@ -16,8 +16,8 @@ protected:
 
 	static int id_throwables;
     const int id_obj;	
-    const int x;
-    const int y;
+    const float x;
+    const float y;
     const float angle_rad;
     const b2Vec2 velocity;
 	const float angular_velocity;
@@ -32,8 +32,8 @@ protected:
 public:
 	Throwable(Stage& stage
 	, Worm* owner
-	, const int x
-	, const int y
+	, const float x
+	, const float y
 	, const float angle_rad
 	, const b2Vec2 velocity
 	, const float angular_velocity
